requirements.txt
Pillow~=5.1.0
tensorflow~=1.13.1
werkzeug~=0.14.0
